requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.20.0
nltk>=3.6.0
rouge-score>=0.1.2
pyyaml>=6.0
av>=10.0.0
huggingface_hub>=0.16.0
bert-score>=0.3.12
pillow>=9.0.0
tqdm>=4.66.0
requests>=2.28.0
logging>=0.5.1.2 